src/test/javascript/spec/app/entities/reponse-demande-de-service/reponse-demande-de-service.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { take, map } from 'rxjs/operators';
import * as moment from 'moment';
import { DATE_TIME_FORMAT } from 'app/shared/constants/input.constants';
import { ReponseDemandeDeServiceService } from 'app/entities/reponse-demande-de-service/reponse-demande-de-service.service';
import { IReponseDemandeDeService, ReponseDemandeDeService } from 'app/shared/model/reponse-demande-de-service.model';

describe('Service Tests', () => {
    describe('ReponseDemandeDeService Service', () => {
        let injector: TestBed;
        let service: ReponseDemandeDeServiceService;
        let httpMock: HttpTestingController;
        let elemDefault: IReponseDemandeDeService;
        let expectedResult;
        let currentDate: moment.Moment;
        beforeEach(() => {
            TestBed.configureTestingModule({
                imports: [HttpClientTestingModule]
            });
            expectedResult = {};
            injector = getTestBed();
            service = injector.get(ReponseDemandeDeServiceService);
            httpMock = injector.get(HttpTestingController);
            currentDate = moment();

            elemDefault = new ReponseDemandeDeService(0, currentDate);
        });

        describe('Service methods', () => {
            it('should find an element', () => {
                const returnedFromService = Object.assign(
                    {
                        dateReponse: currentDate.format(DATE_TIME_FORMAT)
                    },
                    elemDefault
                );
                service
                    .find(123)
                    .pipe(take(1))
                    .subscribe(resp => (expectedResult = resp));

                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(returnedFromService);
                expect(expectedResult).toMatchObject({ body: elemDefault });
            });

            it('should create a ReponseDemandeDeService', () => {
                const returnedFromService = Object.assign(
                    {
                        id: 0,
                        dateReponse: currentDate.format(DATE_TIME_FORMAT)
                    },
                    elemDefault
                );
                const expected = Object.assign(
                    {
                        dateReponse: currentDate
                    },
                    returnedFromService
                );
                service
                    .create(new ReponseDemandeDeService(null))
                    .pipe(take(1))
                    .subscribe(resp => (expectedResult = resp));
                const req = httpMock.expectOne({ method: 'POST' });
                req.flush(returnedFromService);
                expect(expectedResult).toMatchObject({ body: expected });
            });

            it('should update a ReponseDemandeDeService', () => {
                const returnedFromService = Object.assign(
                    {
                        dateReponse: currentDate.format(DATE_TIME_FORMAT)
                    },
                    elemDefault
                );

                const expected = Object.assign(
                    {
                        dateReponse: currentDate
                    },
                    returnedFromService
                );
                service
                    .update(expected)
                    .pipe(take(1))
                    .subscribe(resp => (expectedResult = resp));
                const req = httpMock.expectOne({ method: 'PUT' });
                req.flush(returnedFromService);
                expect(expectedResult).toMatchObject({ body: expected });
            });

            it('should return a list of ReponseDemandeDeService', () => {
                const returnedFromService = Object.assign(
                    {
                        dateReponse: currentDate.format(DATE_TIME_FORMAT)
                    },
                    elemDefault
                );
                const expected = Object.assign(
                    {
                        dateReponse: currentDate
                    },
                    returnedFromService
                );
                service
                    .query(expected)
                    .pipe(
                        take(1),
                        map(resp => resp.body)
                    )
                    .subscribe(body => (expectedResult = body));
                const req = httpMock.expectOne({ method: 'GET' });
                req.flush([returnedFromService]);
                httpMock.verify();
                expect(expectedResult).toContainEqual(expected);
            });

            it('should delete a ReponseDemandeDeService', () => {
                service.delete(123).subscribe(resp => (expectedResult = resp.ok));

                const req = httpMock.expectOne({ method: 'DELETE' });
                req.flush({ status: 200 });
                expect(expectedResult);
            });
        });

        afterEach(() => {
            httpMock.verify();
        });
    });
});
